Refactor ListCrimeActivity to extend from a BaseActivity
* Allows for the new CrimesMapActivity to reuse permissions code
