+++ app/src/main/AndroidManifest.xml
@@ -1,8 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest package="aoemeron.github.io.crimesmvp"
           xmlns:android="http://schemas.android.com/apk/res/android">
+
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
     <uses-permission android:name="android.permission.INTERNET"/>
+
     <application
         android:name="io.github.aoemerson.crimesmvp.application.CrimeApp"
         android:allowBackup="true"
@@ -11,6 +13,13 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name="io.github.aoemerson.crimesmvp.view.ListCrimesActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN"/>
+
+                <category android:name="android.intent.category.LAUNCHER"/>
+            </intent-filter>
+        </activity>
+        <activity android:name="io.github.aoemerson.crimesmvp.view.CrimesMapActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
