+++ app/src/main/java/io/github/aoemerson/crimesmvp/view/ListCrimesActivity.java
@@ -1,12 +1,7 @@
 package io.github.aoemerson.crimesmvp.view;
 
-import android.Manifest;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.support.annotation.StringRes;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.ListView;
 import android.widget.ProgressBar;
@@ -22,14 +17,12 @@ import io.github.aoemerson.crimesmvp.model.data.Crime;
 import io.github.aoemerson.crimesmvp.presenter.CrimeListPresenterImpl;
 import io.github.aoemerson.crimesmvp.presenter.DaggerLocalCrimesComponent;
 
-public class ListCrimesActivity extends AppCompatActivity implements CrimesView {
+public class ListCrimesActivity extends BaseActivity {
 
-    private static final int LOCATION_PERM_REQ_CODE = 1;
     @Inject CrimeListPresenterImpl crimeListPresenter;
     private ProgressBar progressBar;
     private ListView listView;
     private CrimeListViewAdapter crimeListViewAdapter;
-    private LocationPermissionRequestCallback locationRequestCallback;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -68,21 +61,6 @@ public class ListCrimesActivity extends AppCompatActivity implements CrimesView 
     }
 
     @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == LOCATION_PERM_REQ_CODE && locationRequestCallback != null) {
-            boolean granted = false;
-            for (int result : grantResults) {
-                granted = granted || (result == PackageManager.PERMISSION_GRANTED);
-            }
-            if (granted)
-                locationRequestCallback.onLocationPermissionGranted();
-            else
-                locationRequestCallback.onLocationPermissionDenied();
-        }
-    }
-
-    @Override
     public void setCrimes(List<Crime> crimes) {
         crimeListViewAdapter.addAll(crimes);
         crimeListViewAdapter.notifyDataSetChanged();
@@ -104,19 +82,6 @@ public class ListCrimesActivity extends AppCompatActivity implements CrimesView 
     }
 
     @Override
-    public void requestLocationPermission(LocationPermissionRequestCallback callback) {
-        if (!hasLocationPersmission()) {
-            locationRequestCallback = callback;
-            ActivityCompat.requestPermissions(this,
-                    new String[] {Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION}, LOCATION_PERM_REQ_CODE);
-
-        } else {
-            callback.onLocationPermissionGranted();
-        }
-
-    }
-
-    @Override
     public boolean isProgressShown() {
         return progressBar.getVisibility() == View.VISIBLE;
     }
@@ -134,13 +99,6 @@ public class ListCrimesActivity extends AppCompatActivity implements CrimesView 
     @Override
     public void showLocationUnavailableError() {
         showError(R.string.err_location_not_obtained);
-    }
-
-    @Override
-    public boolean hasLocationPersmission() {
-        return ActivityCompat
-                .checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED || ActivityCompat
-                .checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED;
     }
 
     private void showError(@StringRes int errResId) {
